pyproject.toml
[project]
name = "agendadoc-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.127.1",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.30",
]
